[refactor] #5 refactor signin, signup
