+++ src/user/repository/user.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { Prisma, user } from '@prisma/client';
 import { PrismaService } from 'src/prisma.service';
-import { signupDto } from './dto/signup.dto';
+import { signupDto } from '../dto/signup.dto';
 
 @Injectable()
 export class UserRepository {
