--- src/user/user.interface.ts
@@ -1,7 +1,0 @@
-export interface User {
-  user_id: string;
-  email: string;
-  passwd: string;
-  name: string;
-  created_at: Date;
-}
