+++ src/user/user.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserService } from './service/user.service';
 import { PrismaService } from 'src/prisma.service';
-import { UserRepository } from './user.repository';
+import { UserRepository } from './repository/user.repository';
 import { UserEntity } from './user.entities';
 
 @Module({
